// src/app/page/master/pegawai-sk-medical/pegawai-sk-medical.interface.ts
import { EventEmitter } from '@angular/core';

export interface PegawaiSkMedical {
    kode?: any;
    kdKategoriPegawai?: number;
    kdMasaKerja?: string;
    kdJabatanStruktural?: string;
    kdProduk?: string;
    kdKondisiProduk?: string;
    Resume?: string;
    Keterangan?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}